pyproject.toml
[tool.poetry]
name = "scrape_notifier"
version = "0.1.0"
description = ""
authors = ["Jannis Kowalick <jannisko99@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
python-telegram-bot = "^13.13"
toml = "^0.10.2"
SQLAlchemy = "^1.4.39"
alembic = "^1.8.1"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
scrape-notifier = "scrape_notifier.__main__:cli"
